+++ meow-keypad.el
@@ -30,6 +30,7 @@
 ;;
 ;;; Code:
 
+(require 'cl-lib)
 (require 'subr-x)
 (require 'meow-var)
 (require 'meow-util)
@@ -66,9 +67,7 @@
   (let* ((keybind (if meow--keypad-base-keymap
 		      (lookup-key meow--keypad-base-keymap keys)
 		    (key-binding keys))))
-    (unless (and (meow--is-self-insertp keybind)
-		 (not meow-keypad-self-insert-undefined))
-	  keybind)))
+    keybind))
 
 (defun meow--keypad-has-sub-meta-keymap-p ()
   "Check if there's a keymap belongs to Meta prefix.
@@ -408,8 +407,8 @@ If there are beacons, execute it at every beacon."
 (defun meow--keypad-try-execute ()
   "Try execute command, return t when the translation progress can be ended.
 
-If there is a command available on the current key binding,
-try replacing the last modifier and try again."
+This function supports a fallback behavior, where it allows to use `SPC
+x f' to execute `C-x C-f' or `C-x f' when `C-x C-f' is not bound."
   (unless (or meow--use-literal
               meow--use-meta
               meow--use-both)
@@ -439,23 +438,32 @@ try replacing the last modifier and try again."
         (meow--keypad-try-execute))
        (t
         (setq meow--prefix-arg nil)
-        (message "%s is undefined" (meow--keypad-format-keys nil))
         (meow--keypad-quit)
+        (if (or (eq t meow-keypad-leader-transparent)
+                (eq meow--keypad-previous-state meow-keypad-leader-transparent))
+          (let* ((key (meow--parse-input-event last-input-event))
+                 (origin-cmd (cl-some (lambda (m)
+                                        (lookup-key m key))
+                                      (current-active-maps)))
+                 (remapped-cmd (command-remapping origin-cmd))
+                 (cmd-to-call (if (member remapped-cmd '(undefined nil))
+                                  (or origin-cmd 'undefined)
+                                remapped-cmd)))
+            (meow--keypad-execute cmd-to-call))
+          (message "%s is undefined" key-str))
         t)))))
 
 (defun meow--keypad-handle-input-with-keymap (input-event)
   "Handle INPUT-EVENT with `meow-keypad-state-keymap'.
 
 Return t if handling is completed."
-  (if (numberp input-event)
-      (let* ((k (if (= 27 input-event)
-                    [escape]
-                  (kbd (single-key-description input-event))))
-             (cmd (lookup-key meow-keypad-state-keymap k)))
-        (if cmd
-            (call-interactively cmd)
-          (meow--keypad-handle-input-event input-event)))
-    (meow--keypad-quit)))
+  (if (equal 'escape last-input-event)
+      (meow--keypad-quit)
+    (setq last-command-event last-input-event)
+    (let ((kbd (single-key-description input-event)))
+      (if-let* ((cmd (lookup-key meow-keypad-state-keymap (read-kbd-macro kbd))))
+          (call-interactively cmd)
+        (meow--keypad-handle-input-event input-event)))))
 
 (defun meow--keypad-handle-input-event (input-event)
   "Handle the INPUT-EVENT.